+++ packages/appkit/exports/react.ts
@@ -70,3 +70,13 @@ export function useAppKitBalance() {
 }
 
 export { useAppKitAccount } from '@reown/appkit-controllers/react'
+
+export {
+  AppKitButton,
+  AppKitNetworkButton,
+  AppKitConnectButton,
+  AppKitAccountButton
+} from '../src/library/react/components.js'
+
+export { AppKitProvider } from '../src/library/react/providers.js'
+export type { AppKitProviderProps } from '../src/library/react/providers.js'