+++ packages/controllers/exports/react.ts
@@ -12,7 +12,11 @@ import { ConnectorController } from '../src/controllers/ConnectorController.js'
 import { OptionsController } from '../src/controllers/OptionsController.js'
 import { ConnectionControllerUtil } from '../src/utils/ConnectionControllerUtil.js'
 import { CoreHelperUtil } from '../src/utils/CoreHelperUtil.js'
-import type { UseAppKitAccountReturn, UseAppKitNetworkReturn } from '../src/utils/TypeUtil.js'
+import type {
+  NamespaceTypeMap,
+  UseAppKitAccountReturn,
+  UseAppKitNetworkReturn
+} from '../src/utils/TypeUtil.js'
 import { AssetUtil, StorageUtil } from './utils.js'
 
 // -- Types ------------------------------------------------------------
@@ -80,8 +84,13 @@ export function useAppKitAccount(options?: { namespace?: ChainNamespace }): UseA
   const activeConnectorId = activeConnectorIds[chainNamespace]
   const connections = ConnectionController.getConnections(chainNamespace)
   const allAccounts = connections.flatMap(connection =>
-    connection.accounts.map(({ address }) =>
-      CoreHelperUtil.createAccount(chainNamespace, address, 'eoa')
+    connection.accounts.map(({ address, type, publicKey }) =>
+      CoreHelperUtil.createAccount(
+        chainNamespace,
+        address,
+        (type || 'eoa') as NamespaceTypeMap[ChainNamespace],
+        publicKey
+      )
     )
   )