+++ packages/controllers/src/controllers/SwapController.ts
@@ -1,7 +1,7 @@
 import { proxy, subscribe as sub } from 'valtio/vanilla'
 import { subscribeKey as subKey } from 'valtio/vanilla/utils'
 
-import { type Address, type Hex, NumberUtil } from '@reown/appkit-common'
+import { type Address, type CaipNetworkId, type Hex, NumberUtil } from '@reown/appkit-common'
 import { ConstantsUtil as CommonConstantsUtil } from '@reown/appkit-common'
 import { W3mFrameRpcConstants } from '@reown/appkit-wallet/utils'
 
@@ -49,12 +49,12 @@ type TransactionParams = {
 }
 
 class TransactionError extends Error {
-  shortMessage?: string
+  displayMessage?: string
 
-  constructor(message?: string, shortMessage?: string) {
+  constructor(message?: string, displayMessage?: string) {
     super(message)
     this.name = 'TransactionError'
-    this.shortMessage = shortMessage
+    this.displayMessage = displayMessage
   }
 }
 
@@ -92,7 +92,9 @@ export interface SwapControllerState {
   slippage: number
 
   // Tokens
+  caipNetworkId?: CaipNetworkId
   tokens?: SwapTokenWithBalance[]
+  tokensLoading?: boolean
   suggestedTokens?: SwapTokenWithBalance[]
   popularTokens?: SwapTokenWithBalance[]
   foundTokens?: SwapTokenWithBalance[]
@@ -171,7 +173,7 @@ const initialState: SwapControllerState = {
   providerFee: undefined
 }
 
-const state = proxy<SwapControllerState>(initialState)
+const state = proxy<SwapControllerState>({ ...initialState })
 
 // -- Controller ---------------------------------------- //
 const controller = {
@@ -252,9 +254,7 @@ const controller = {
   },
 
   setToTokenAmount(amount: string) {
-    state.toTokenAmount = amount
-      ? NumberUtil.formatNumberToLocalString(amount, TO_AMOUNT_DECIMALS)
-      : ''
+    state.toTokenAmount = amount ? NumberUtil.toFixed(amount, TO_AMOUNT_DECIMALS) : ''
   },
 
   async setTokenPrice(address: string, target: SwapInputTarget) {
@@ -302,6 +302,7 @@ const controller = {
     state.myTokensWithBalance = initialState.myTokensWithBalance
     state.tokensPriceMap = initialState.tokensPriceMap
     state.initialized = initialState.initialized
+    state.initializing = initialState.initializing
     state.sourceToken = initialState.sourceToken
     state.sourceTokenAmount = initialState.sourceTokenAmount
     state.sourceTokenPriceInUSD = initialState.sourceTokenPriceInUSD
@@ -312,7 +313,6 @@ const controller = {
     state.networkTokenSymbol = initialState.networkTokenSymbol
     state.networkBalanceInUSD = initialState.networkBalanceInUSD
     state.inputError = initialState.inputError
-    state.myTokensWithBalance = initialState.myTokensWithBalance
   },
 
   resetValues() {
@@ -353,40 +353,55 @@ const controller = {
   async fetchTokens() {
     const { networkAddress } = SwapController.getParams()
 
-    await SwapController.getTokenList()
     await SwapController.getNetworkTokenPrice()
     await SwapController.getMyTokensWithBalance()
 
-    const networkToken = state.tokens?.find(token => token.address === networkAddress)
+    const networkToken = state.myTokensWithBalance?.find(token => token.address === networkAddress)
 
     if (networkToken) {
       state.networkTokenSymbol = networkToken.symbol
       SwapController.setSourceToken(networkToken)
-      SwapController.setSourceTokenAmount('1')
+      SwapController.setSourceTokenAmount('0')
     }
   },
 
   async getTokenList() {
-    const tokens = await SwapApiUtil.getTokenList()
+    const activeCaipNetworkId = ChainController.state.activeCaipNetwork?.caipNetworkId
 
-    state.tokens = tokens
-    state.popularTokens = tokens.sort((aTokenInfo, bTokenInfo) => {
-      if (aTokenInfo.symbol < bTokenInfo.symbol) {
-        return -1
-      }
-      if (aTokenInfo.symbol > bTokenInfo.symbol) {
-        return 1
-      }
+    if (state.caipNetworkId === activeCaipNetworkId && state.tokens) {
+      return
+    }
 
-      return 0
-    })
-    state.suggestedTokens = tokens.filter(token => {
-      if (ConstantsUtil.SWAP_SUGGESTED_TOKENS.includes(token.symbol)) {
-        return true
-      }
+    try {
+      state.tokensLoading = true
+      const tokens = await SwapApiUtil.getTokenList(activeCaipNetworkId)
 
-      return false
-    }, {})
+      state.tokens = tokens
+      state.caipNetworkId = activeCaipNetworkId
+      state.popularTokens = tokens.sort((aTokenInfo, bTokenInfo) => {
+        if (aTokenInfo.symbol < bTokenInfo.symbol) {
+          return -1
+        }
+        if (aTokenInfo.symbol > bTokenInfo.symbol) {
+          return 1
+        }
+
+        return 0
+      })
+      state.suggestedTokens = tokens.filter(token => {
+        if (ConstantsUtil.SWAP_SUGGESTED_TOKENS.includes(token.symbol)) {
+          return true
+        }
+
+        return false
+      })
+    } catch (error) {
+      state.tokens = []
+      state.popularTokens = []
+      state.suggestedTokens = []
+    } finally {
+      state.tokensLoading = false
+    }
   },
 
   async getAddressPrice(address: string) {
@@ -534,8 +549,8 @@ const controller = {
       if (!quoteToAmount) {
         AlertController.open(
           {
-            shortMessage: 'Incorrect amount',
-            longMessage: 'Please enter a valid amount'
+            displayMessage: 'Incorrect amount',
+            debugMessage: 'Please enter a valid amount'
           },
           'error'
         )
@@ -760,14 +775,14 @@ const controller = {
       state.loadingApprovalTransaction = false
     } catch (err) {
       const error = err as TransactionError
-      state.transactionError = error?.shortMessage as unknown as string
+      state.transactionError = error?.displayMessage as unknown as string
       state.loadingApprovalTransaction = false
-      SnackController.showError(error?.shortMessage || 'Transaction error')
+      SnackController.showError(error?.displayMessage || 'Transaction error')
       EventsController.sendEvent({
         type: 'track',
         event: 'SWAP_APPROVAL_ERROR',
         properties: {
-          message: error?.shortMessage || error?.message || 'Unknown',
+          message: error?.displayMessage || error?.message || 'Unknown',
           network: ChainController.state.activeCaipNetwork?.caipNetworkId || '',
           swapFromToken: SwapController.state.sourceToken?.symbol || '',
           swapToToken: SwapController.state.toToken?.symbol || '',
@@ -844,14 +859,14 @@ const controller = {
       return transactionHash
     } catch (err) {
       const error = err as TransactionError
-      state.transactionError = error?.shortMessage
+      state.transactionError = error?.displayMessage
       state.loadingTransaction = false
-      SnackController.showError(error?.shortMessage || 'Transaction error')
+      SnackController.showError(error?.displayMessage || 'Transaction error')
       EventsController.sendEvent({
         type: 'track',
         event: 'SWAP_ERROR',
         properties: {
-          message: error?.shortMessage || error?.message || 'Unknown',
+          message: error?.displayMessage || error?.message || 'Unknown',
           network: ChainController.state.activeCaipNetwork?.caipNetworkId || '',
           swapFromToken: SwapController.state.sourceToken?.symbol || '',
           swapToToken: SwapController.state.toToken?.symbol || '',