+++ packages/controllers/tests/controllers/AlertController.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest'
+import { afterEach, describe, expect, it, vi } from 'vitest'
 
 import { AlertController, OptionsController } from '../../exports/index.js'
 
@@ -7,8 +7,12 @@ OptionsController.state.debug = true
 
 // -- Tests --------------------------------------------------------------------
 describe('AlertController', () => {
+  afterEach(() => {
+    vi.restoreAllMocks()
+  })
+
   it('should show info state', () => {
-    AlertController.open({ shortMessage: 'Info message' }, 'info')
+    AlertController.open({ displayMessage: 'Info message' }, 'info')
     expect(AlertController.state).toStrictEqual({
       message: 'Info message',
       variant: 'info',
@@ -17,7 +21,7 @@ describe('AlertController', () => {
   })
 
   it('should show success state', () => {
-    AlertController.open({ shortMessage: 'Success message' }, 'success')
+    AlertController.open({ displayMessage: 'Success message' }, 'success')
     expect(AlertController.state).toStrictEqual({
       message: 'Success message',
       variant: 'success',
@@ -26,7 +30,7 @@ describe('AlertController', () => {
   })
 
   it('should show warning state', () => {
-    AlertController.open({ shortMessage: 'Warning message' }, 'warning')
+    AlertController.open({ displayMessage: 'Warning message' }, 'warning')
     expect(AlertController.state).toStrictEqual({
       message: 'Warning message',
       variant: 'warning',
@@ -35,7 +39,7 @@ describe('AlertController', () => {
   })
 
   it('should show error state', () => {
-    AlertController.open({ shortMessage: 'Error message' }, 'error')
+    AlertController.open({ displayMessage: 'Error message' }, 'error')
     expect(AlertController.state).toStrictEqual({
       message: 'Error message',
       variant: 'error',
@@ -50,5 +54,22 @@ describe('AlertController', () => {
       variant: 'info',
       open: false
     })
+  })
+
+  it('should show error code', () => {
+    const consoleErrorSpy = vi.spyOn(console, 'error').mockImplementation(() => {})
+    AlertController.open(
+      {
+        code: 'APKT005',
+        displayMessage: 'Unverified Domain',
+        debugMessage:
+          'Embedded wallet load failed. Ensure your domain is verified in https://dashboard.reown.com.'
+      },
+      'error'
+    )
+    expect(consoleErrorSpy).toHaveBeenCalledWith(
+      'Embedded wallet load failed. Ensure your domain is verified in https://dashboard.reown.com.',
+      { code: 'APKT005' }
+    )
   })
 })